remove duplicate vite plugin
