+++ .changeset/config.json
@@ -7,9 +7,7 @@
 			"@openpolicy/sdk",
 			"@openpolicy/core",
 			"@openpolicy/vite",
-			"@openpolicy/vite-auto-collect",
 			"@openpolicy/cli",
-			"@openpolicy/astro",
 			"@openpolicy/react",
 			"@openpolicy/vue",
 			"@openpolicy/renderers"
