+++ apps/docs/content/docs/generation/vite.md
@@ -1,11 +1,11 @@
 ---
 title: Vite
-description: Generate policy files at build time with the OpenPolicy Vite plugin
+description: Scan source files for data-collection annotations at build time
 ---
 
 See the [Generation Overview](/generation/overview) for context.
 
-The `@openpolicy/vite` plugin generates policy files during the Vite build and watches for config changes in dev.
+The `@openpolicy/vite` plugin scans your source for `collecting()` and `thirdParty()` calls and populates the `@openpolicy/sdk` auto-collected registry so `dataCollected` / `thirdParties` are available at runtime. It does not write policy files — render policies directly in your app via SDK components, or use the CLI for static output.
 
 ## Install
 
@@ -27,39 +27,24 @@ export default defineConfig({
 });
 ```
 
-By default this reads `openpolicy.ts` from the project root and writes output to `public/policies/` as Markdown.
+By default the plugin walks `src/` for `.ts` / `.tsx` files and refreshes its registry on change during `vite dev`.
 
 ## Options
 
 ```ts
 openPolicy({
-  configPath: "openpolicy.ts",   // path to your config file
-  formats: ["markdown"],         // "markdown" | "html" | "pdf"
-  outDir: "public/policies",     // output directory
+  srcDir: "src",
+  extensions: [".ts", ".tsx"],
+  ignore: [],
+  thirdParties: { usePackageJson: false },
 })
 ```
 
 | Option | Type | Default | Description |
 |---|---|---|---|
-| `configPath` | `string` | `"openpolicy.ts"` | Path to your config file |
-| `formats` | `OutputFormat[]` | `["markdown"]` | One or more output formats |
-| `outDir` | `string` | `"public/policies"` | Directory to write files into |
+| `srcDir` | `string` | `"src"` | Directory walked for `collecting()` / `thirdParty()` calls |
+| `extensions` | `string[]` | `[".ts", ".tsx"]` | File extensions scanned |
+| `ignore` | `string[]` | `[]` | Extra directory names to skip (appended to built-in ignores) |
+| `thirdParties.usePackageJson` | `boolean` | `false` | Auto-detect third parties from `package.json` against the known-packages registry |
 
-## With auto-collect
-
-If you use `autoCollect()` to scan source files for data collection annotations, add it before `openPolicy()` in your plugin array so the scan completes before the policy config is imported.
-
-```ts
-// vite.config.ts
-import { autoCollect } from "@openpolicy/vite-auto-collect";
-import { openPolicy } from "@openpolicy/vite";
-
-export default defineConfig({
-  plugins: [
-    autoCollect(),
-    openPolicy(),
-  ],
-});
-```
-
-See [Auto-collect](/policies/auto-collect) for full setup and options.
+See [Auto-collect](/policies/auto-collect) for annotation patterns and examples.
