+++ apps/www/package.json
@@ -24,11 +24,9 @@
 	},
 	"devDependencies": {
 		"@types/react": "^19.0.0",
-		"@openpolicy/astro": "workspace:*",
 		"@openpolicy/core": "workspace:*",
 		"@openpolicy/react": "workspace:*",
 		"@openpolicy/sdk": "workspace:*",
-		"@openpolicy/vite": "workspace:*",
 		"@tailwindcss/typography": "^0.5.19",
 		"shadcn": "latest"
 	}
