+++ apps/www/src/content/frameworks/astro.md
@@ -1,6 +1,6 @@
 ---
 title: OpenPolicy with Astro
-description: Generate and render privacy and cookie policies in Astro sites using the OpenPolicy Astro integration.
+description: Render privacy and cookie policies directly in Astro pages using @openpolicy/core.
 framework: Astro
 icon: simple-icons:astro
 pubDate: 2026-03-23
@@ -8,26 +8,12 @@ pubDate: 2026-03-23
 
 # OpenPolicy with Astro
 
-The `@openpolicy/astro` integration wires up the Vite plugin for you and lets you import compiled policy HTML directly into `.astro` components.
+Compile policies directly in your Astro page frontmatter with `@openpolicy/core` — no integration, no generated files.
 
 ## Install
 
 ```bash
-bun add @openpolicy/sdk @openpolicy/astro
-```
-
-## Configure the integration
-
-```ts
-// astro.config.mjs
-import { defineConfig } from "astro/config";
-import openPolicy from "@openpolicy/astro";
-
-export default defineConfig({
-  integrations: [
-    openPolicy({ configs: ["./src/openpolicy.ts"] }),
-  ],
-});
+bun add @openpolicy/sdk @openpolicy/core @openpolicy/renderers
 ```
 
 ## Define your policy
@@ -48,7 +34,15 @@ export default definePrivacyPolicy({
 ```astro
 ---
 // src/pages/privacy-policy.astro
-import policy from "../privacy-policy.html?raw";
+import { compile, expandOpenPolicyConfig } from "@openpolicy/core";
+import { renderHTML } from "@openpolicy/renderers";
+import openpolicy from "../openpolicy";
+
+const policies = expandOpenPolicyConfig(openpolicy);
+const privacyPolicy = policies.find((p) => p.type === "privacy");
+if (!privacyPolicy) throw new Error("Privacy policy not found");
+
+const policy = renderHTML(compile(privacyPolicy));
 ---
 
 <html lang="en">
@@ -61,7 +55,7 @@ import policy from "../privacy-policy.html?raw";
 
 ## Why OpenPolicy for Astro
 
-- **Zero JS by default** — the policy compiles to static HTML; no client bundle added
+- **Zero JS by default** — the policy compiles to static HTML in frontmatter; no client bundle added
 - **`set:html` directive** — Astro's built-in way to render trusted HTML strings
-- **Seamless Vite integration** — the Astro integration handles plugin wiring for you
+- **No integration required** — just call `compile()` in frontmatter where you need it
 - **Ships with your content** — policy lives alongside your Markdown and MDX content, version controlled in git
