+++ apps/www/src/pages/index.astro
@@ -115,7 +115,7 @@ const frameworks = [
     </p>
     <CodeBlock code={instrumentTs} lang="ts" filename="src/actions/create-user.ts" />
     <p class="text-sm text-gray-500 mt-6 mb-2">
-      Wire <code>autoCollect()</code> into your Vite config so the scan runs on every build.
+      Wire <code>openPolicy()</code> into your Vite config so the scan runs on every build.
     </p>
     <CodeBlock code={viteCfg} lang="ts" filename="vite.config.ts" />
   </section>
