+++ examples/tanstack/package.json
@@ -26,7 +26,7 @@
 	"devDependencies": {
 		"@playwright/test": "^1.53.2",
 		"@openpolicy/sdk": "workspace:*",
-		"@openpolicy/vite-auto-collect": "workspace:*",
+		"@openpolicy/vite": "workspace:*",
 		"@tailwindcss/vite": "^4.2.1",
 		"@types/node": "^25.3.5",
 		"@types/react": "^19.2.14",
