+++ examples/tanstack/vite.config.ts
@@ -1,4 +1,4 @@
-import { autoCollect } from "@openpolicy/vite-auto-collect";
+import { openPolicy } from "@openpolicy/vite";
 import tailwindcss from "@tailwindcss/vite";
 import { tanstackStart } from "@tanstack/react-start/plugin/vite";
 import viteReact from "@vitejs/plugin-react";
@@ -14,7 +14,7 @@ export default defineConfig({
 	plugins: [
 		tailwindcss(),
 		tsConfigPaths(),
-		autoCollect({ srcDir: "./src", thirdParties: { usePackageJson: true } }),
+		openPolicy({ srcDir: "./src", thirdParties: { usePackageJson: true } }),
 		tanstackStart(),
 		nitro(),
 		// react's vite plugin must come after start's vite plugin
