+++ package.json
@@ -8,7 +8,6 @@
 		"@changesets/cli": "^2.30.0",
 		"@openpolicy/react": "workspace:*",
 		"@openpolicy/sdk": "workspace:*",
-		"@openpolicy/vite-auto-collect": "workspace:*",
 		"@tanstack/intent": "^0.0.29",
 		"@types/bun": "latest",
 		"@types/node": "^25.5.0",
