+++ packages/sdk/skills/annotate-third-parties/SKILL.md
@@ -2,7 +2,7 @@
 name: annotate-third-parties
 description: >
   Mark third-party service dependencies in source code using thirdParty() so
-  the autoCollect() Vite plugin can populate the thirdParties sentinel for the
+  the openPolicy() Vite plugin can populate the thirdParties sentinel for the
   privacy policy config. Covers manual annotation, usePackageJson auto-detection
   from the KNOWN_PACKAGES registry, combining both approaches, and the Providers
   preset as a static alternative.
@@ -13,14 +13,14 @@ requires:
   - openpolicy/vite-setup
 sources:
   - jamiedavenport/openpolicy:packages/sdk/src/third-parties.ts
-  - jamiedavenport/openpolicy:packages/vite-auto-collect/src/known-packages.ts
+  - jamiedavenport/openpolicy:packages/vite/src/known-packages.ts
 ---
 
 This skill builds on openpolicy/vite-setup. Read it first.
 
 ## How It Works
 
-`thirdParty()` is a **runtime no-op**. It is a build-time marker — `autoCollect()` scans source files statically via `oxc-parser` and extracts the three string literal arguments. The plugin then populates the `thirdParties` sentinel exported from `@openpolicy/sdk`. Without the plugin, `thirdParty()` calls have no effect whatsoever.
+`thirdParty()` is a **runtime no-op**. It is a build-time marker — `openPolicy()` scans source files statically via `oxc-parser` and extracts the three string literal arguments. The plugin then populates the `thirdParties` sentinel exported from `@openpolicy/sdk`. Without the plugin, `thirdParty()` calls have no effect whatsoever.
 
 The populated sentinel must be spread into the top-level `thirdParties` field in the config file. Without the spread, the plugin output is discarded and the policy renders with no third parties listed.
 
@@ -32,11 +32,11 @@ Three moving parts must all be in place:
 
 ```ts
 import { defineConfig } from "vite";
-import { autoCollect } from "@openpolicy/vite-auto-collect";
+import { openPolicy } from "@openpolicy/vite";
 
 export default defineConfig({
   plugins: [
-    autoCollect({
+    openPolicy({
       thirdParties: { usePackageJson: true }, // optional — see below
     }),
   ],
@@ -92,13 +92,13 @@ Deduplication is by `name` — the first occurrence wins when files are walked i
 
 ### 2. usePackageJson auto-detection
 
-`autoCollect({ thirdParties: { usePackageJson: true } })` reads `package.json` at build time and cross-references every entry in `dependencies` and `devDependencies` against the built-in KNOWN_PACKAGES registry. Matched packages are merged into `thirdParties` automatically, with deduplication by service name.
+`openPolicy({ thirdParties: { usePackageJson: true } })` reads `package.json` at build time and cross-references every entry in `dependencies` and `devDependencies` against the built-in KNOWN_PACKAGES registry. Matched packages are merged into `thirdParties` automatically, with deduplication by service name.
 
 This is the fastest way to cover common services (Stripe, Sentry, PostHog, Vercel, etc.) without any source annotations.
 
 ```ts
 // vite.config.ts
-autoCollect({ thirdParties: { usePackageJson: true } })
+openPolicy({ thirdParties: { usePackageJson: true } })
 ```
 
 See [references/known-packages.md](references/known-packages.md) for the full list of ~30 auto-detected packages.
@@ -113,7 +113,7 @@ Use this combination when:
 
 ```ts
 // vite.config.ts
-autoCollect({ thirdParties: { usePackageJson: true } })
+openPolicy({ thirdParties: { usePackageJson: true } })
 
 // src/lib/custom-analytics.ts
 import { thirdParty } from "@openpolicy/sdk";
@@ -156,22 +156,22 @@ thirdParties: [...thirdParties, Providers.Cloudflare],
 
 ## Common Mistakes
 
-### HIGH — Using thirdParty() without the autoCollect() plugin and expecting a runtime effect
+### HIGH — Using thirdParty() without the openPolicy() plugin and expecting a runtime effect
 
-`thirdParty()` is defined as a no-op function that immediately returns `undefined`. It has no runtime behaviour. The only way it produces output in the policy is when `autoCollect()` is present in `vite.config.ts` and scans the file at build time.
+`thirdParty()` is defined as a no-op function that immediately returns `undefined`. It has no runtime behaviour. The only way it produces output in the policy is when `openPolicy()` is present in `vite.config.ts` and scans the file at build time.
 
 ```ts
 // WRONG: calling thirdParty() and expecting it to affect the policy at runtime
 import { thirdParty } from "@openpolicy/sdk";
 thirdParty("Stripe", "Payment processing", "https://stripe.com/privacy");
-// Without autoCollect() in vite.config.ts, this does nothing
+// Without openPolicy() in vite.config.ts, this does nothing
 ```
 
 ```ts
-// Correct: autoCollect() in vite.config.ts is the required prerequisite
+// Correct: openPolicy() in vite.config.ts is the required prerequisite
 // vite.config.ts
-import { autoCollect } from "@openpolicy/vite-auto-collect";
-export default defineConfig({ plugins: [autoCollect()] });
+import { openPolicy } from "@openpolicy/vite";
+export default defineConfig({ plugins: [openPolicy()] });
 
 // src/lib/stripe.ts — scanned at build time by the plugin
 import { thirdParty } from "@openpolicy/sdk";
@@ -180,7 +180,7 @@ thirdParty("Stripe", "Payment processing", "https://stripe.com/privacy");
 
 ### HIGH — Not spreading thirdParties sentinel into thirdParties
 
-Even when `autoCollect()` scans the source and populates the sentinel, the populated value is discarded unless it is imported and spread into the top-level `thirdParties` field in the config. The privacy policy then renders with an empty third-party section, which may be legally invalid.
+Even when `openPolicy()` scans the source and populates the sentinel, the populated value is discarded unless it is imported and spread into the top-level `thirdParties` field in the config. The privacy policy then renders with an empty third-party section, which may be legally invalid.
 
 ```ts
 // WRONG: thirdParties not imported or spread
@@ -206,12 +206,12 @@ The KNOWN_PACKAGES registry covers ~30 common npm packages. Without `usePackageJ
 
 ```ts
 // WRONG: usePackageJson not enabled — known packages not auto-detected
-autoCollect()
+openPolicy()
 ```
 
 ```ts
 // Correct
-autoCollect({ thirdParties: { usePackageJson: true } })
+openPolicy({ thirdParties: { usePackageJson: true } })
 ```
 
 ## References
@@ -220,6 +220,6 @@ autoCollect({ thirdParties: { usePackageJson: true } })
 - `packages/sdk/src/third-parties.ts` — thirdParty() implementation (runtime no-op)
 - `packages/sdk/src/providers.ts` — Providers preset objects
 - `packages/sdk/src/auto-collected.ts` — thirdParties sentinel definition
-- `packages/vite-auto-collect/src/index.ts` — autoCollect() plugin, AutoCollectOptions type
-- `packages/vite-auto-collect/src/analyse.ts` — static AST extraction logic
-- `packages/vite-auto-collect/src/known-packages.ts` — KNOWN_PACKAGES registry source
+- `packages/vite/src/index.ts` — openPolicy() plugin, OpenPolicyOptions type
+- `packages/vite/src/analyse.ts` — static AST extraction logic
+- `packages/vite/src/known-packages.ts` — KNOWN_PACKAGES registry source
