+++ packages/sdk/src/auto-collected.ts
@@ -1,8 +1,8 @@
 /**
- * Placeholder populated by `@openpolicy/vite-auto-collect` during a Vite
+ * Placeholder populated by `@openpolicy/vite` during a Vite
  * build. The plugin intercepts this module's resolution and replaces it with
  * the scanned categories, so the literal default below is only used as a
- * fallback when no auto-collect plugin is active — in which case spreading
+ * fallback when the plugin is not active — in which case spreading
  * it into `dataCollected` is a no-op.
  *
  * @example
@@ -22,7 +22,7 @@
 export const dataCollected: Record<string, string[]> = {};
 
 /**
- * Placeholder populated by `@openpolicy/vite-auto-collect` during a Vite
+ * Placeholder populated by `@openpolicy/vite` during a Vite
  * build. The plugin intercepts this module's resolution and replaces it with
  * the third-party services discovered via `thirdParty()` calls, so the literal
  * default below is only used as a fallback when no auto-collect plugin is
