--- packages/vite-auto-collect/package.json
@@ -1,45 +1,0 @@
-{
-	"name": "@openpolicy/vite-auto-collect",
-	"version": "0.0.23",
-	"type": "module",
-	"description": "Vite plugin that scans source files for @openpolicy/sdk collecting() calls and populates autoCollected() at build time",
-	"license": "GPL-3.0-only",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/jamiedavenport/openpolicy",
-		"directory": "packages/vite-auto-collect"
-	},
-	"keywords": [
-		"vite-plugin",
-		"tanstack-intent"
-	],
-	"files": [
-		"dist",
-		"README.md",
-		"skills"
-	],
-	"exports": {
-		".": {
-			"import": "./dist/index.js",
-			"types": "./dist/index.d.ts"
-		}
-	},
-	"scripts": {
-		"dev": "rolldown --watch -c",
-		"build": "rolldown -c",
-		"check-types": "tsc --noEmit",
-		"test": "bun test"
-	},
-	"peerDependencies": {
-		"vite": ">=4.0.0"
-	},
-	"dependencies": {
-		"oxc-parser": "^0.124.0"
-	},
-	"devDependencies": {
-		"@openpolicy/sdk": "workspace:*",
-		"@openpolicy/tooling": "workspace:*",
-		"vite": "catalog:",
-		"@tanstack/intent": "^0.0.29"
-	}
-}
