--- packages/vite-auto-collect/rolldown.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "rolldown";
-import { dts } from "rolldown-plugin-dts";
-
-export default defineConfig({
-	input: "./src/index.ts",
-	plugins: [dts()],
-	platform: "node",
-	external: (id) => /^node:/.test(id) || /^[^./]/.test(id),
-	output: { format: "esm", dir: "dist" },
-});
