--- packages/vite-auto-collect/tsconfig.json
@@ -1,7 +1,0 @@
-{
-	"extends": "@openpolicy/tooling/base",
-	"include": ["src"],
-	"compilerOptions": {
-		"outDir": "dist"
-	}
-}
