+++ packages/vite/rolldown.config.ts
@@ -5,9 +5,6 @@ export default defineConfig({
 	input: "./src/index.ts",
 	plugins: [dts()],
 	platform: "node",
-	external: (id) =>
-		id !== "@openpolicy/core" &&
-		id !== "@openpolicy/renderers" &&
-		(/^node:/.test(id) || /^[^./]/.test(id)),
+	external: (id) => /^node:/.test(id) || /^[^./]/.test(id),
 	output: { format: "esm", dir: "dist" },
 });
