+++ packages/vite/src/known-packages.ts
