+++ packages/vite/src/scan.test.ts
