+++ packages/vite/src/scan.ts
